+++ packages/aws-lambda-rest/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @mondrian-framework/aws-lambda-rest
 
+## 1.0.4
+
+### Patch Changes
+
+- Improve performance
+- Updated dependencies
+  - @mondrian-framework/module@1.0.4
+  - @mondrian-framework/model@1.0.4
+  - @mondrian-framework/utils@1.0.4
+  - @mondrian-framework/rest@1.0.6
+
 ## 1.0.3
 
 ### Patch Changes