+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mondrian-framework/utils
 
+## 1.0.4
+
+### Patch Changes
+
+- Improve performance
+
 ## 1.0.3
 
 ### Patch Changes